requirements.txt
opencv-python
pillow
numpy
scikit-learn